Revert "Revert "Start to refactor "tabs" link helpers""
